requirements.txt
Django==1.8
argparse==1.2.1
dj-database-url==0.2.1
dj-static==0.0.5
django-crispy-forms==1.4.0
django-toolbelt==0.0.1
gunicorn==0.17.4
jedi==0.7.0
newrelic==2.50.0.39
psycopg2==2.5
static==0.4
wsgiref==0.1.2
raven==5.0.0
requests==2.2.1
reportlab==3.0
django-filter-0.10.0
djangorestframework-3.1.3
markdown-2.6.2
